Initial import.
Over the years, I had done modifications to TeX-9. This is the first
time they get thrown in github. Among them, are the snippet <leader>Z to
use \includeonly, with omni-completion, and the generic insertion of an
environment now leaves you in insert mode, in the first (empty) line
after the \begin line.