A LOT of things. See below...
First update the license with the data from yours truly here.

Then remove everything that has to do with compilation, because (repeat
after me) Vim is not an IDE!!!!

Also remove stuff to synctec, because that is heavily dependent on the
pdf viewer used.

Removed the skeletons, because I have my own (see my repo LaTeX.git).

Simplified the insertion of snippets: only for basic environment, with
<Localleader>B. All other things are done with snipMate (to be forked).

The config leader is either localleader, if it is set, or ':' (the
default). The value of leader is no longer used.

The changes for mappings in had in my vim settings are now in the
plugin (author's privilege).

But I deleted greek/math mappings for bibtex. I don't think it makes
much sense...