Make gd great again!
In TeX-9's original design, if you hit gd over a \ref key, you go to the
corresponding label. If you hit gb over a \cite key, you go to the
corresponding bib entry. Now for this to work, the cursor had to be over
the "key" part, when the mapping was invoked. Some time ago I changed
this, so that the cursor could be anywhere \ref{key} or \cite{key} (or
whatever variations of those commands where used). However, there was a
problem: my strategy did not work if the cursor was over the '\' char!

I fixed that creating a function that checked where the char below the
cursor was '\', and if so, move one char to the right---and mapped this
latter function to gd. I had to do the same thing for gb, which
suggested the idea of using just one function for both: and detecting
from the string that got selected whether I was dealing with a \ref or a
\cite. This is done in function tex_seven#QueryMap.

Additionally, there are some extra caveats when dealing with \cite; see
that function's comments for details.

There was also the following bug: when searching for a \label, the
preview window was *always* opened, even if the search failed; this is
now fixed.